feat: add floating as link
