+++ packages/manager/apps/pci-instances/package.json
@@ -16,11 +16,11 @@
     "type:check": "tsc --noEmit"
   },
   "dependencies": {
-    "@ovh-ux/manager-config": "^8.0.0",
-    "@ovh-ux/manager-core-api": "^0.9.0",
-    "@ovh-ux/manager-pci-common": "^0.14.4",
-    "@ovh-ux/manager-react-components": "^1.43.0",
-    "@ovh-ux/manager-react-core-application": "^0.11.5",
+    "@ovh-ux/manager-config": "^8.0.2",
+    "@ovh-ux/manager-core-api": "^0.10.0",
+    "@ovh-ux/manager-pci-common": "^0.15.2",
+    "@ovh-ux/manager-react-components": "^1.45.0",
+    "@ovh-ux/manager-react-core-application": "^0.11.9",
     "@ovh-ux/manager-react-shell-client": "^0.8.5",
     "@ovh-ux/manager-tailwind-config": "^0.2.1",
     "@ovhcloud/ods-common-core": "17.2.2",
@@ -42,7 +42,7 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.7.0",
-    "@ovh-ux/manager-vite-config": "^0.9.0",
+    "@ovh-ux/manager-vite-config": "^0.9.1",
     "@tanstack/react-query-devtools": "^5.51.21",
     "@testing-library/dom": "^9.3.4",
     "@testing-library/jest-dom": "^6.4.2",
@@ -61,8 +61,8 @@
     "postcss": "^8.4.35",
     "tailwindcss": "^3.4.4",
     "uuid": "^9.0.1",
-    "vite": "^5.2.13",
-    "vitest": "^1.2.0"
+    "vite": "^5.3.3",
+    "vitest": "^2.1.9"
   },
   "regions": [
     "CA",
