+++ packages/manager/apps/pci-instances/src/data/hooks/instance/useInstances.ts
@@ -23,7 +23,7 @@ export type TUseInstancesQueryParams = DeepReadonly<{
   filters: FilterWithLabel[];
 }>;
 
-export type TAddressType = 'public' | 'private';
+export type TAddressType = 'public' | 'private' | 'floating';
 
 export type TInstanceStatusSeverity = 'success' | 'error' | 'warning' | 'info';
 export type TInstanceStatusState = TInstanceStatusDto;
