+++ packages/manager/apps/pci-instances/src/pages/instances/datagrid/cell/ListCell.spec.tsx
@@ -11,13 +11,18 @@ const items: TListCellItem[] = [
     id: '78910',
     name: 'bar',
   },
+  { id: '3612', name: 'floatingIp', href: 'floaint' },
 ];
 
 describe('Considering the ListCell component', () => {
   test('Should render the component with given items', () => {
     render(<ListCell isLoading={false} items={items} />);
     items.forEach((item) => {
-      expect(screen.getByText(item.name)).toBeInTheDocument();
+      const el = screen.getByText(item.name);
+      expect(el).toBeInTheDocument();
+      if (item.href) {
+        expect(el).toHaveAttribute('href', item.href);
+      }
     });
   });
 });
